Open436-SectionService/requirements.txt
# Django 核心
Django==4.2.7
djangorestframework==3.14.0

# 数据库
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# API 文档
drf-spectacular==0.26.5

# CORS 支持
django-cors-headers==4.3.1

# 环境变量
python-dotenv==1.0.0

# 服务发现
python-consul2==0.1.5

# HTTP 客户端（调用其他服务）
requests==2.31.0

# 日志增强
python-json-logger==2.0.7

# 时间处理
python-dateutil==2.8.2

# 验证工具
validators==0.22.0

# 生产环境服务器
gunicorn==21.2.0

